requirements.txt
PySide6>=6.6
twelvelabs>=0.2.0
python-dotenv>=1.0
opentimelineio
